requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
SpeechRecognition==3.10.0
pyttsx3==2.90
python-dateutil==2.8.2
pytz==2023.3
python-dotenv==1.0.0
pyaudio==0.2.13
